requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
Faker==30.8.0
Flask==2.2.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.3
greenlet==3.1.1
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
lorem==0.1.1
Mako==1.3.6
MarkupSafe==3.0.2
mysqlclient==2.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==2.2.3
